[OWL-1866] Move new health API to another file.
